Refactor: made changes suggested in PR
