+++ src/pages/index.js
@@ -1,0 +1,5 @@
+export { AuthorizationPage } from "./authorizationPage/authorizationPage";
+export { CartPage } from "./CartPage/CartPage";
+export { LandingPage } from "./landingPage/LandingPage";
+export { ProductsListingPage } from "./productsListingPage/productListing";
+export { WishListPage } from "./wishlistPage/WishListPage";
